Now moving last file from binlogproto to gorpcbinlogcommon.
